+++ onmt/trainer.py
@@ -1,3 +1,4 @@
+
 """
     This is the loadable seq2seq trainer library that is
     in charge of training details, loss compute, and statistics.
@@ -13,8 +14,13 @@ import torch
 import traceback
 import onmt.utils
 from onmt.utils.logging import logger
+<<<<<<< HEAD
 from onmt.model_builder import load_test_model
 from onmt.utils.lm_prior_loss import lm_prior_loss
+=======
+from onmt.translate.utils import ScoringPreparator
+from onmt.scorers import get_scorers_cls, build_scorers
+>>>>>>> 173e113b1220b32c1d03a0f8d3f9d44759c071de
 
 
 def build_trainer(opt, device_id, model, fields, optim, model_saver=None):
@@ -37,6 +43,12 @@ def build_trainer(opt, device_id, model, fields, optim, model_saver=None):
     valid_loss = onmt.utils.loss.build_loss_compute(
         model, tgt_field, opt, train=False)
 
+    scoring_preparator = ScoringPreparator(fields, opt)
+    scorers_cls = get_scorers_cls(opt.train_metrics)
+    train_scorers = build_scorers(opt, scorers_cls)
+    scorers_cls = get_scorers_cls(opt.valid_metrics)
+    valid_scorers = build_scorers(opt, scorers_cls)
+
     trunc_size = opt.truncated_decoder  # Badly named...
     shard_size = opt.max_generator_batches if opt.model_dtype == 'fp32' else 0
     norm_method = opt.normalization
@@ -46,6 +58,7 @@ def build_trainer(opt, device_id, model, fields, optim, model_saver=None):
     average_decay = opt.average_decay
     average_every = opt.average_every
     dropout = opt.dropout
+    attention_dropout = opt.attention_dropout
     dropout_steps = opt.dropout_steps
     if device_id >= 0:
         gpu_rank = opt.gpu_ranks[device_id]
@@ -59,6 +72,7 @@ def build_trainer(opt, device_id, model, fields, optim, model_saver=None):
         if opt.early_stopping > 0 else None
 
     report_manager = onmt.utils.build_report_manager(opt, gpu_rank)
+<<<<<<< HEAD
     if opt.lm_prior_model:
         opt.gpu = 0
         opt.fp32 = False
@@ -77,10 +91,16 @@ def build_trainer(opt, device_id, model, fields, optim, model_saver=None):
         vocab = tgt_field.vocab
 
     trainer = onmt.Trainer(model, train_loss, valid_loss, optim, trunc_size,
+=======
+    trainer = onmt.Trainer(model,
+                           train_loss, valid_loss,
+                           scoring_preparator, train_scorers, valid_scorers,
+                           optim, trunc_size,
+>>>>>>> 173e113b1220b32c1d03a0f8d3f9d44759c071de
                            shard_size, norm_method,
                            accum_count, accum_steps,
-                           n_gpu, gpu_rank,
-                           gpu_verbose_level, report_manager,
+                           n_gpu, gpu_rank, gpu_verbose_level,
+                           opt.train_eval_steps, report_manager,
                            with_align=True if opt.lambda_align > 0 else False,
                            model_saver=model_saver if gpu_rank <= 0 else None,
                            average_decay=average_decay,
@@ -88,18 +108,22 @@ def build_trainer(opt, device_id, model, fields, optim, model_saver=None):
                            model_dtype=opt.model_dtype,
                            earlystopper=earlystopper,
                            dropout=dropout,
+<<<<<<< HEAD
                            dropout_steps=dropout_steps,
                            lm_prior_model=lm_prior_model,
                            vocab=vocab,
                            lm_prior_lambda=lm_prior_lambda,
                            lm_prior_tau=lm_prior_tau)
+=======
+                           attention_dropout=attention_dropout,
+                           dropout_steps=dropout_steps)
+>>>>>>> 173e113b1220b32c1d03a0f8d3f9d44759c071de
     return trainer
 
 
 class Trainer(object):
     """
     Class that controls the training process.
-
     Args:
             model(:py:class:`onmt.models.model.NMTModel`): translation model
                 to train
@@ -122,20 +146,33 @@ class Trainer(object):
                 Thus nothing will be saved if this parameter is None
     """
 
-    def __init__(self, model, train_loss, valid_loss, optim,
+    def __init__(self, model, train_loss, valid_loss,
+                 scoring_preparator, train_scorers, valid_scorers,
+                 optim,
                  trunc_size=0, shard_size=32,
                  norm_method="sents", accum_count=[1],
                  accum_steps=[0],
                  n_gpu=1, gpu_rank=1, gpu_verbose_level=0,
+                 train_eval_steps=200,
                  report_manager=None, with_align=False, model_saver=None,
                  average_decay=0, average_every=1, model_dtype='fp32',
+<<<<<<< HEAD
                  earlystopper=None, dropout=[0.3], dropout_steps=[0],
                  lm_prior_model=None, vocab=None,
                  lm_prior_lambda=None, lm_prior_tau=None):
+=======
+                 earlystopper=None, dropout=[0.3], attention_dropout=[0.1],
+                 dropout_steps=[0]):
+>>>>>>> 173e113b1220b32c1d03a0f8d3f9d44759c071de
         # Basic attributes.
+
         self.model = model
         self.train_loss = train_loss
         self.valid_loss = valid_loss
+
+        self.scoring_preparator = scoring_preparator
+        self.train_scorers = train_scorers
+        self.valid_scorers = valid_scorers
         self.optim = optim
         self.trunc_size = trunc_size
         self.shard_size = shard_size
@@ -147,6 +184,7 @@ class Trainer(object):
         self.gpu_rank = gpu_rank
         self.gpu_verbose_level = gpu_verbose_level
         self.report_manager = report_manager
+        self.train_eval_steps = train_eval_steps
         self.with_align = with_align
         self.model_saver = model_saver
         self.average_decay = average_decay
@@ -155,6 +193,7 @@ class Trainer(object):
         self.model_dtype = model_dtype
         self.earlystopper = earlystopper
         self.dropout = dropout
+        self.attention_dropout = attention_dropout
         self.dropout_steps = dropout_steps
         self.lm_prior_model = lm_prior_model
         self.vocab = vocab
@@ -171,6 +210,16 @@ class Trainer(object):
         # Set model in training mode.
         self.model.train()
 
+    def training_eval_handler(self, scorer, batch, mode="train"):
+        """Trigger metrics calculations"""
+        preds, texts_ref = self.scoring_preparator.translate(
+            model=self.model,
+            batch=batch,
+            gpu_rank=self.gpu_rank,
+            step=self.optim.training_step,
+            mode=mode)
+        return scorer.compute_score(preds, texts_ref)
+
     def _accum_count(self, step):
         for i in range(len(self.accum_steps)):
             if step > self.accum_steps[i]:
@@ -180,9 +229,11 @@ class Trainer(object):
     def _maybe_update_dropout(self, step):
         for i in range(len(self.dropout_steps)):
             if step > 1 and step == self.dropout_steps[i] + 1:
-                self.model.update_dropout(self.dropout[i])
-                logger.info("Updated dropout to %f from step %d"
-                            % (self.dropout[i], step))
+                self.model.update_dropout(self.dropout[i],
+                                          self.attention_dropout[i])
+                logger.info("Updated dropout/attn dropout to %f %f at step %d"
+                            % (self.dropout[i],
+                               self.attention_dropout[i], step))
 
     def _accum_batches(self, iterator):
         batches = []
@@ -235,7 +286,6 @@ class Trainer(object):
               iterations.
             valid_iter: A generator that returns the next validation batch.
             valid_steps: Run evaluation every this many iterations.
-
         Returns:
             The gathered statistics.
         """
@@ -279,16 +329,14 @@ class Trainer(object):
                 self.optim.learning_rate(),
                 report_stats)
 
-            if valid_iter is not None and step % valid_steps == 0:
+            if (valid_iter is not None and step % valid_steps == 0 and
+                    self.gpu_rank == 0):
                 if self.gpu_verbose_level > 0:
                     logger.info('GpuRank %d: validate step %d'
                                 % (self.gpu_rank, step))
+
                 valid_stats = self.validate(
                     valid_iter, moving_average=self.moving_average)
-                if self.gpu_verbose_level > 0:
-                    logger.info('GpuRank %d: gather valid stat \
-                                step %d' % (self.gpu_rank, step))
-                valid_stats = self._maybe_gather_stats(valid_stats)
                 if self.gpu_verbose_level > 0:
                     logger.info('GpuRank %d: report stat step %d'
                                 % (self.gpu_rank, step))
@@ -299,6 +347,7 @@ class Trainer(object):
                     self.earlystopper(valid_stats, step)
                     # If the patience has reached the limit, stop training
                     if self.earlystopper.has_stopped():
+                        logger.info("earlystopper has_stopped!")
                         break
 
             if (self.model_saver is not None
@@ -349,8 +398,33 @@ class Trainer(object):
                     # Compute loss.
                     _, batch_stats = self.valid_loss(batch, outputs, attns)
 
+                    stats.update(batch_stats)
+
+            # Compute validation metrics (at batch.dataset level)
+            computed_metrics = {}
+            for i, metric in enumerate(self.valid_scorers):
+                logger.info("UPDATING VALIDATION {}".format(metric))
+                self.valid_scorers[
+                    metric]["value"] = self.training_eval_handler(
+                        scorer=self.valid_scorers[metric]["scorer"],
+                        batch=batch,
+                        mode="valid")
+                computed_metrics[
+                    metric] = self.valid_scorers[metric]["value"]
+                logger.info(
+                    "validation {}: {}".format(
+                        metric, self.valid_scorers[metric]["value"])
+                        )
+                # Compute stats
+                batch_stats = onmt.utils.Statistics(
+                    batch_stats.loss,
+                    batch_stats.n_words,
+                    batch_stats.n_correct,
+                    computed_metrics)
+
                 # Update statistics.
                 stats.update(batch_stats)
+
         if moving_average:
             for param_data, param in zip(model_params_data,
                                          self.model.parameters()):
@@ -390,21 +464,45 @@ class Trainer(object):
                 if self.accum_count == 1:
                     self.optim.zero_grad()
 
-                with torch.cuda.amp.autocast(enabled=self.optim.amp):
-                    outputs, attns = self.model(
-                        src, tgt, src_lengths, bptt=bptt,
-                        with_align=self.with_align)
-                    bptt = True
+                try:
+                    with torch.cuda.amp.autocast(enabled=self.optim.amp):
+                        outputs, attns = self.model(
+                            src, tgt, src_lengths, bptt=bptt,
+                            with_align=self.with_align)
+                        bptt = True
 
-                    # 3. Compute loss.
-                    loss, batch_stats = self.train_loss(
-                        batch,
-                        outputs,
-                        attns,
-                        normalization=normalization,
-                        shard_size=self.shard_size,
-                        trunc_start=j,
-                        trunc_size=trunc_size)
+                        # 3. Compute loss.
+                        loss, batch_stats = self.train_loss(
+                            batch,
+                            outputs,
+                            attns,
+                            normalization=normalization,
+                            shard_size=self.shard_size,
+                            trunc_start=j,
+                            trunc_size=trunc_size)
+
+                    step = self.optim.training_step
+                    if (
+                            step % self.train_eval_steps == 0 and
+                            self.n_gpu > 0
+                    ):
+                        # Compute and save stats
+                        computed_metrics = {}
+                        for i, metric in enumerate(self.train_scorers):
+                            logger.info("UPDATING TRAINING {}".format(metric))
+                            self.train_scorers[
+                                metric]["value"] = self.training_eval_handler(
+                                scorer=self.train_scorers[
+                                    metric]["scorer"],
+                                batch=batch,
+                                mode="train")
+                            logger.info(
+                                "training {}: {}".format(
+                                    metric, self.train_scorers[
+                                        metric]["value"]))
+                            computed_metrics[
+                                metric] = self.train_scorers[metric]["value"]
+                        batch_stats.computed_metrics = computed_metrics
 
                     if self.lm_prior_model is not None:
                         lm_loss = lm_prior_loss(self.model,
@@ -414,7 +512,6 @@ class Trainer(object):
                     else:
                         lm_loss = 0
 
-                try:
                     if loss is not None:
                         total_loss = loss + self.lm_prior_lambda * lm_loss
                         self.optim.backward(total_loss)
@@ -422,10 +519,14 @@ class Trainer(object):
                     total_stats.update(batch_stats)
                     report_stats.update(batch_stats)
 
-                except Exception:
-                    traceback.print_exc()
-                    logger.info("At step %d, we removed a batch - accum %d",
-                                self.optim.training_step, k)
+                except Exception as exc:
+                    trace_content = traceback.format_exc()
+                    if "CUDA out of memory" in trace_content:
+                        logger.info("Step %d, cuda OOM - batch removed",
+                                    self.optim.training_step)
+                    else:
+                        traceback.print_exc()
+                        raise exc
 
                 # 4. Update the parameters and statistics.
                 if self.accum_count == 1: