+++ onmt/translate/beam_search.py
@@ -189,18 +189,18 @@ class BeamSearchBase(DecodeStrategy):
                     self.is_finished[i].all()
             else:
                 finish_flag = self.top_beam_finished[i] != 0
-            if finish_flag and len(self.hypotheses[b]) >= self.n_best:
+            if finish_flag and len(self.hypotheses[b]) >= self.beam_size:
                 best_hyp = sorted(
-                    self.hypotheses[b], key=lambda x: x[0], reverse=True)
+                    self.hypotheses[b], key=lambda x: x[0],
+                    reverse=True)[:self.n_best]
                 for n, (score, pred, attn) in enumerate(best_hyp):
-                    if n >= self.n_best:
-                        break
                     self.scores[b].append(score)
                     self.predictions[b].append(pred)  # ``(batch, n_best,)``
                     self.attention[b].append(
                         attn if attn is not None else [])
             else:
                 non_finished_batch.append(i)
+
         non_finished = torch.tensor(non_finished_batch)
         # If all sentences are translated, no need to go further.
         if len(non_finished) == 0:
@@ -433,15 +433,9 @@ class GNMTGlobalScorer(object):
         # these warnings indicate that either the alpha/beta
         # forces a penalty to be a no-op, or a penalty is a no-op but
         # the alpha/beta would suggest otherwise.
-        if length_penalty is None or length_penalty == "none":
-            if alpha != 0:
-                warnings.warn("Non-default `alpha` with no length penalty. "
-                              "`alpha` has no effect.")
-        else:
-            # using some length penalty
-            if length_penalty == "wu" and alpha == 0.:
-                warnings.warn("Using length penalty Wu with alpha==0 "
-                              "is equivalent to using length penalty none.")
+        if length_penalty is not None and alpha == 0.:
+            warnings.warn("Using length penalty with alpha==0 "
+                          "is equivalent to using length penalty none.")
         if coverage_penalty is None or coverage_penalty == "none":
             if beta != 0:
                 warnings.warn("Non-default `beta` with no coverage penalty. "